Revert "Feature/email export"
