+++ android/app/build.gradle
@@ -143,7 +143,6 @@ dependencies {
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
-    compile project(':RNMail')
 }
 
 // Run this once to be able to run the application with BUCK
