+++ android/app/src/main/java/com/comedycompanion/MainApplication.java
@@ -12,7 +12,6 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.chirag.RNMail.*;
 
 import java.util.Arrays;
 import java.util.List;
@@ -29,12 +28,11 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-          new VectorIconsPackage(),
-          new KCKeepAwakePackage(),
-          new RNFSPackage(),
-          new RNSoundPackage(),
-          new ReactNativeAudioPackage(),
-          new RNMail()
+            new VectorIconsPackage(),
+            new KCKeepAwakePackage(),
+            new RNFSPackage(),
+            new RNSoundPackage(),
+            new ReactNativeAudioPackage()
       );
     }
   };
