+++ android/settings.gradle
@@ -9,7 +9,5 @@ include ':react-native-sound'
 project(':react-native-sound').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-sound/android')
 include ':react-native-audio'
 project(':react-native-audio').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-audio/android')
-include ':RNMail', ':app'
-project(':RNMail').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-mail/android')
 
 include ':app'
