+++ package.json
@@ -7,6 +7,8 @@
     "test": "jest"
   },
   "dependencies": {
+    "react-native-sliding-panes": "1.0.1",
+    "react-native-sweet-record": "1.0.0",
     "debounce": "^1.0.2",
     "moment": "2.18.1",
     "react": "15.4.2",
@@ -16,13 +18,10 @@
     "react-native-hamburger": "0.0.2",
     "react-native-icloudstore": "^0.9.0",
     "react-native-keep-awake": "^2.0.4",
-    "react-native-mail": "git://github.com/dereksweet/react-native-mail.git",
     "react-native-material-design-searchbar": "^1.1.2",
     "react-native-radio-buttons": "^0.14.0",
-    "react-native-sliding-panes": "1.0.1",
     "react-native-sortable-listview": "^0.2.2",
     "react-native-sound": "^0.10.3",
-    "react-native-sweet-record": "1.0.0",
     "react-native-swipeout": "^2.1.3",
     "react-native-ui-xg": "0.0.2",
     "react-native-vector-icons": "^4.0.0",
