+++ src/actions/routingActions.js
@@ -69,12 +69,6 @@ export function toggleAbout() {
   }
 }
 
-export function toggleDownload() {
-  return {
-    type: types.TOGGLE_DOWNLOAD
-  }
-}
-
 export function hideLoading() {
   return {
     type: types.HIDE_LOADING
