+++ src/models/setting.js
@@ -7,8 +7,6 @@ const defaults = {
   _set_lists_sort_order: 'DESC',
   _shows_sort_field: '_date',
   _shows_sort_order: 'DESC',
-  _export_email: '',
-  _export_email_type: 'formatted',
   _about_version_seen: 0.0
 };
 
@@ -29,8 +27,6 @@ export default class Setting extends SweetModel {
     _set_lists_sort_order: defaults._set_lists_sort_order,
     _shows_sort_field: defaults._shows_sort_field,
     _shows_sort_order: defaults._shows_sort_order,
-    _export_email: defaults._export_email,
-    _export_email_type: defaults._export_email_type,
     _about_version_seen: defaults._about_version_seen
   })
   {
@@ -43,8 +39,6 @@ export default class Setting extends SweetModel {
     this._set_lists_sort_order = data._set_lists_sort_order || defaults._set_lists_sort_order;
     this._shows_sort_field     = data._shows_sort_field || defaults._shows_sort_field;
     this._shows_sort_order     = data._shows_sort_order || defaults._shows_sort_order;
-    this._export_email         = data._export_email || defaults._export_email;
-    this._export_email_type    = data._export_email_type || defaults._export_email_type;
     this._about_version_seen   = data._about_version_seen || defaults._about_version_seen;
   }
 }
