--- src/stylesheets/downloadStyles.js
@@ -1,13 +1,0 @@
-import { StyleSheet, Platform } from 'react-native';
-
-export default downloadStyles = StyleSheet.create({
-  exportEmailInput: {
-    flex: 1,
-    height: 40,
-    borderColor: 'gray',
-    borderWidth: 1,
-    paddingLeft: 10,
-    paddingRight: 10,
-    backgroundColor: '#FFFFFF'
-  }
-});
